docs website (#1228)
* add docosaurus default documentation website

* fix port number

seems docusaurus has a bug and you can't pass in PORT via command-line
because they misinterpret this as an integer when it really should be a
string. however, they handle the port correctly when passed as --port
directly to their app.

* prettier modification

* commit docs page build

I suppose it makes sense to commit the build directory for the new docs
website. The reason to commit it is that it is a static website and we
want to be able to direct an nginx instance directly to this folder.
Otherwise we will have to add yet another build process. Then again
maybe we can do exactly that... we're going to build moneybutton anyway,
maybe we can just build it then.

* commit prettified file again

* remove build directory

* ignore these files

* new config. ignore file.

* set up pages for each of the important things to document

Content not yet written.

* primitive react-money-button documentation

* primitive html docs