clients and webhooks (#1321)
* use ownerId to look up webhook

* switch from "ownerId" to "appId"

* update docs to reference appKey instead of ownerId

I'm using twitter as a reference. On twitter, you must make new apps,
and each app has a name, description, requested permissins, callback
URL, and finally a "consumer key" which must be used to reference that
app.

https://apps.twitter.com/

* new database schema

this commit breaks the UI.

TODO: Fix the UI.

* "apps"

This is a notion of apps created by a users. The apps "own" buttons. The
apps are how we keep track of webhook URLs and webhook secrets.

* apps -> clients

* fix comment

* make client secret the same

* minor

* fix integration tests

* consumerKey -> clientIdentifier

The names "client secret" and "client identifier" are what's used in the
oauth spec so we'll use the same names here. Note that "client
identifier" is oddly similar to "client id" but they are NOT the same.
"client identifer" is an id specifically for the owner of the client as
well as all other users to see publicly, where as "client id" really
just an internal id for us in our own database.

* attempt to fix travis error

* remove webhook stuff from the UI

It used to be that webhooks are a user setting. But now they are on the
clients. So the UI is broken. The next step is to add some UI for
editing the clients which allows setting the webhook URL for each
client.

* display clients in the UI

* update payments API to use correct webhook in client

* emilio's concern: findByClientIdentifier->getByClientIdentifier

* fix emilio's concern over logging unawaited promise error

* emilio's concern: allow clientSecret to be null with no default

* emilio's comment: remove wiggly fawn client secret from migration

* emilio's comment: allow clientSecret to be null with no default

* emilio's comment: fix documentation

* fix test

* fix test