webhooks (#1427)
* refactor input processing

We currently are detecting double spends and malleability but aren't yet
actually calling the webhook when these events occur. This commit
refactors the input code to have fewer indendent blocks and adds many
comments explaining what's going on, along with a mock webhook call. I
also added some comments to the UTXO model.

* ownerId -> clientIdentifier

Some time recently we decided that we should use OAuth for
authentication and this comes with a notion of "clients" which are the
apps that allow users to authenticate. We also recently decided that
each "app" that a developer might make should map to these clients, and
that therefore the proper place to specify a webhook URL is in the
client. This *replaces* the previous notion of an "owner" of an app.

So this commit fixes two things:
- Replace all instances of ownerId with clientIdentifier
- Add a notion of a webhookSecret that is sent back to the webhook

The clientIdentifier, confusingly, is the public-facing identifer for a
client, which is *not* equal to clientId. The clientId is our own
internal id for a client that is used in our own database. Yes, this is
confusing. We think it's better to keep our own database internally
consistent with the way we use ids and for the clientIdentifier to be a
different thing. See the OAuth spec.

Furthermore we don't want to use the clientSecret for the webhook
because that's not really what it's for. So I've created a new
"webhookSecret" that is generated randomly and sent back through the
webhook. The webhookSecret is a way for developers to know that calls to
their webhook are legit, because it is not public knowledge. It is only
known by our servers and their servers and is never seen in the wild.

* better name

crypto.generateCryptographicallySecureRandomString()

->

crypto.secureRandomString()

* slightly better naming for code readability

* abstract pingWebhook into own utility

I've created a pingWebhook utility in the API subproject (that seems
like the most appropriate place, as the webhook is a part of our API).
This method utility is called from both inside the monitor and inside
the payments API.

Question: Does this mean we call the webhook twice? If we broadcast a tx
to bitcore, does it give the same tx back to us via the socket API? What
about Bitcoin ABC - if we switch to zeromq, and we give it a tx, will it
give the same tx back to us via the zeromq interface?

Note that as of this commit, the answer is no, because we are not yet
pinging the webhook from the monitor on the first seen of a tx in the
monitor. So this is a question to be aware of as we continue to improve
the webhooks.

* fix tests

* fix tests

* wrap pingWebhook in try/catch

As Emilio points out, since this method is intended to be used without
an await most of the time, it's best to wrap the contents in a
try/catch since it won't get caught anywhere else.

* depend on api

* fix one broken test

* export pingWebhook from mb-api

* use constructor convention for mocks

* ping webhook for confirmed transactions

* use a database transaction

We will run multiple monitors at the same time. If two monitors process
the same confirmed tx at the same time, there is a chance of a race
condition where they both read the same payment from the database,
change its status to confirmed, and then ping the webhook, meaning we
ping their webhook twice. This can be fixed using database transactions.
We merge the read and write into one operation with a transaction
(database transaction, not bitcoin transaction :) ), and then only ping
the webhook if the status changed. Only one of the two monitors will be
able to update the status, and so the one that doesn't update it does
not ping the webhook, therefore we only ping the webhook once per tx
confirm no matter how many monitors we are running.

* a test for processTransaction

* test for confirmed txs

* test for processConfirmedTransaction

* add comment

* we're never going to send an error over the webhook

* this should be transacting too

* simpler logic & fewer queries

* more efficient not to fetch again

* comment typo

* confirm webhook url is valid before calling it

* better punctuation

* enforce http urls

* fix bug

* don't send the "amount". add outputs to payment.

* remove obsolete comment

* refactor for fewer tabs

* js standard nit

* put outer try/catch in new method

* js standard nit

* comment wording

* import logger

* woops, this too

* API naming consistency

* handle blockHeight = 0

* also get payment, don't attempt to rebuild

* better to retrieve from db again

* normalizedTxid

* fix logic

* another try at the logic

* better names