docs: fix sidebars and header (#2989)
* docs: fix sidebars and header

The new OAuth pages were not listed on the front-page or header. There
actually wasn't enough room on mobile for a new section, so it seems to
make the most sense to merge all API docs together into a group. It's
not like there is more than one API - they are all aspects of the same
thing.

* spelling and capitalization